Approval/unapproval events correctly record feature review snapshot
Previously the snapshots relied on there being comment.body in the
lira event payload, but this was not working as not all lira events
have the same structure (and hence comments are at different node points
and are sometimes an array of comments).

This commit changes the functionality to depend on the jira key alone
(which is always present) and uses this to look up a ticket and subsequently
the associated feature reviews.

Signed-off-by: Dennis Ideler <dennis.ideler@fundingcircle.com>
Signed-off-by: Neil Atkinson <neil@new-bamboo.co.uk>