Feature reviews know whether they are approved.
They know this based on the whether the approved_at is set, which is set
in the feature review repository when events are processed.
Previously, we were wrapping a feature review in a FeatureReviewWithStatuses 
decorator.

This change also requires moving the updating of pull requests to the 
feature review repository, because (now) the feature_reviews do not undergo
a state change until the jira event is processed by the feature review
repository and an approved_at is recorded in the snapshot.