Make ticket spec more agnostic of what approval states are
We should not need to update the spec every time we add a new approval
state as it does not give us any increased confidence.
Instead we should be testing that #approved? works for any given
approval state.

Minor changes to approved_statuses for readability and sorting states
into a more logical order (has no effect on the functionality).