Move generation of GitHub URLs to GitRepositoryLocation
Previously it was being done in the FeatureReviewWithStatuses and the
ReleasesController. How to generate GitHub URLs is not a concern of
either of them. GitRepositoryLocations should know "where" a repository
is (i.e. the location of a repository). This can include where it is on
GitHub. The code is now also more DRY and no longer depends on passing
around repository locations just for generating URLs.

Note that the URLs are not generated on initialization of
a FeatureReviewWithStatuses. This is because it can be a lazy operation.
Not every FeatureReviewWithStatuses will need GitHub URLs.